fixes  Fix 1 — queue_item_id uint32 → uint64:   - proto.hxx: QueueTrackRef, RendererState, QueueTrack, MsgQueueTracksRemoved   - proto.cxx: all decode sites (no more truncating casts), encode uses writeUint64Field   - qcmgr.hxx/cxx: all vectors, function signatures, loop variables   - mpdctl.hxx/cxx: item map and lookup functions   - wsession.hxx: callback signature
  Fix 2 — Immediate StateUpdated after SetState:
  - wsession.cxx: after calling on_set_state callback, immediately builds an ack state reflecting the commanded playing_state/position/queue_item and sends it
  via buildStateUpdated() — matches qonductor's synchronous response pattern

  Fix 3 — Heartbeat with actual state:
  - wsession.hxx: added m_last_state + m_state_mutex
  - wsession.cxx: reportState() caches state; sendHeartbeat() sends cached state instead of dummy zeros

  Bonus — DeviceInfo software_version:
  - proto.cxx: encodeDeviceInfo() now writes field 8 (software_version)

+++ src/qconnect/mpdctl.cxx
@@ -301,14 +301,14 @@ void MpdCtl::setStateCallback(MpdStateCallback cb) {
 
 // ---- Queue item mapping -----------------------------------------------------
 
-int MpdCtl::queueItemToMpdId(uint32_t queue_item_id) const {
+int MpdCtl::queueItemToMpdId(uint64_t queue_item_id) const {
     std::lock_guard<std::mutex> lk(m_map_mutex);
     for (const auto& p : m_item_map)
         if (p.first == queue_item_id) return p.second;
     return -1;
 }
 
-void MpdCtl::registerQueueItem(uint32_t queue_item_id, int mpd_id) {
+void MpdCtl::registerQueueItem(uint64_t queue_item_id, int mpd_id) {
     std::lock_guard<std::mutex> lk(m_map_mutex);
     m_item_map.emplace_back(queue_item_id, mpd_id);
 }
